Add a --target-soc option to set_boost_charge.
This needed changes to the session code as it was previously
checking if the car was plugged in, but for this case that
isn't a requirment.
set_boost_charge.py --target-soc will now query a Leaf to
find the current SOC, work out how much power is required
and when is cheapest to apply it, then configure the boost
timer accordingly.